+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to the "export-dir2file" extension will be documented in thi
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [0.0.9] - Jul 2, 2024
+
+- bug fixes
+
+## [0.0.7] - Jul 2, 2024
+
+- Configuration for custom descriptions for export files and active tabs commands
+- fixed issue with duplicate tree structure
+- tree structure now includes on active tabs export
+- minor fixes
+
+## [0.0.6] - Jul 1, 2024
+
+- Rewrite to classes and improve performance
+
+## [0.0.4] - Jun 27, 2024
 
 - Initial release